cleanup and add contrast to UI for readability
