+++ lib/src/media_player/audio_player_handler.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/widgets.dart';
+//import 'package:flutter/widgets.dart';
 import 'package:just_audio/just_audio.dart';
 import 'package:audio_service/audio_service.dart';
